Use class selector instead of element selectors
As far as I can tell `css-modules` locally scope class selectors but not
element selectors in from the component css file. This means that the
`GlobalSelectors` example is slightly misleading since:

```
.root :global p {
  ...
}
```

and

```
.root p {
  ...
}
```

Both have the desired result of styling all the `p` tags inside the
elements with the `.root` class applied.

Updating the example to use a, semantically meaningless, class makes
things clearer for people going through the examples.